move filesize_to_human_readable() to utils/files.py
